+++ app/src/main/res/layout/activity_signin.xml
@@ -9,7 +9,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.example.william.stapp.SignInActivity">
+    tools:context="com.example.william.stapp.SignIn.SignInActivity">
        <!-- Insert Logo Below -->
     <ImageView
         android:id="@+id/logo"
